# swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/register":{"post":{"operationId":"AuthController_signup","summary":"유저 등록 API","description":"유저 등록하는 API 입니다.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"유저등록 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegisteredResponse"}}}}},"tags":["[1] 인증 (Auth): 완료"]}},"/auth/login":{"post":{"operationId":"AuthController_login","summary":"로그인 API","description":"로그인하는 API 입니다.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDto"}}}},"responses":{"200":{"description":"로그인 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginedResponse"}}}}},"tags":["[1] 인증 (Auth): 완료"]}},"/api/users/{id}":{"patch":{"operationId":"UserController_update","summary":"유저 수정 API","description":"유저 수정하는 API 입니다.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"유저 수정 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[2] 유저 (User): 완료"],"security":[{"bearer":[]}]}},"/api/deliverys":{"post":{"operationId":"DeliveryController_createDelivery","summary":"일반 결제 API","description":"\n일반 결제하는 API 입니다.\n<h1>절차</h1>\n<p>\n사전결제 생성 -> 일반 결제 -> 정보 등록 (일반결제 API)\n<p>\n<h1>일반결제 개발 가이드</h1>\n<p>\n<a href=\"https://docs.iamport.kr/implementation/delivery\">아임포트 일반결제 연동</a>\n<a href=\"https://github.com/iamport/iamport-manual/blob/master/%EC%9D%B8%EC%A6%9D%EA%B2%B0%EC%A0%9C/sample/uplus.md\">토스 일반결제 연동</a>\n<a href=\"https://github.com/iamport/iamport-manual/blob/master/%EC%9D%B8%EC%A6%9D%EA%B2%B0%EC%A0%9C/sample/kakao.md\">카카오페이 일반결제 연동</a>\n<a href=\"https://github.com/iamport/iamport-manual/blob/master/%EC%9D%B8%EC%A6%9D%EA%B2%B0%EC%A0%9C/sample/kakao.md\">네이버페이 일반결제 연동</a>\n</p>\n","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDeliveryDto"}}}},"responses":{"200":{"description":"일반 결제 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryCreateResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[5] 결제 (Delivery): 완료"],"security":[{"bearer":[]}]}},"/api/deliverys/card":{"post":{"operationId":"DeliveryController_createCardDelivery","summary":"카드 결제 API","description":"카드 결제하는 API 입니다.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCardDeliveryDto"}}}},"responses":{"200":{"description":"카드 결제 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryCreateResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[5] 결제 (Delivery): 완료"],"security":[{"bearer":[]}]}},"/api/deliverys/bought":{"get":{"operationId":"DeliveryController_getBoughtDeliverys","summary":"구매한 결제내역 리스트 조회 API","description":"구매한 결제내역 리스트 조회하는 API 입니다.","parameters":[],"responses":{"200":{"description":"구매한 결제내역 리스트 조회 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryListResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[5] 결제 (Delivery): 완료"],"security":[{"bearer":[]}]}},"/api/deliverys/bought/{id}":{"get":{"operationId":"DeliveryController_getBoughtDelivery","summary":"구매한 결제내역 상세 조회 API","description":"구매한 결제내역 상세 조회하는 API 입니다.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"구매한 결제내역 상세 조회 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryRetrieveResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[5] 결제 (Delivery): 완료"],"security":[{"bearer":[]}]}},"/api/deliverys/sold":{"get":{"operationId":"DeliveryController_getSoldDeliverys","summary":"판매한 결제내역 상세 조회 API","description":"판매한 결제내역 상세 조회하는 API 입니다.","parameters":[],"responses":{"200":{"description":"판매한 결제내역 리스트 조회 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryRetrieveResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[5] 결제 (Delivery): 완료"],"security":[{"bearer":[]}]}},"/api/deliverys/sold/{id}":{"get":{"operationId":"DeliveryController_getSoldDelivery","summary":"판매한 결제내역 상세 조회 API","description":"판매한 결제내역 상세 조회하는 API 입니다.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"판매한 결제 상세 조회 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryRetrieveResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[5] 결제 (Delivery): 완료"],"security":[{"bearer":[]}]}},"/api/deliverys/prepares":{"post":{"operationId":"DeliveryPrepareController_createPrepare","summary":"사전결제 생성 API","description":"사전결제을 생성하는 API 입니다.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDeliveryPrepareDto"}}}},"responses":{"200":{"description":"사전결제 생성 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryPrepareCreateResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[4] 사전결제 (DeliveryPrepare): 완료"],"security":[{"bearer":[]}]},"get":{"operationId":"DeliveryPrepareController_getDeliveryPrepares","summary":"사전결제 리스트 조회 API","description":"사전결제 리스트를 조회하는 API 입니다.","parameters":[],"responses":{"200":{"description":"사전결제 리스트 조회 성공시 결과 예시","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/Paginated"},{"properties":{"statusCode":{"type":"number","default":200},"message":{"type":"string","default":"리스트 조회를 완료하였습니다"},"data":{"type":"array","items":{"$ref":"#/components/schemas/DeliveryPrepareEntity"}}}}]}}}},"401":{"description":"Unauthorized"}},"tags":["[4] 사전결제 (DeliveryPrepare): 완료"],"security":[{"bearer":[]}]}},"/api/deliverys/prepares/{id}":{"get":{"operationId":"DeliveryPrepareController_getDeliveryPrepare","summary":"사전결제 상세 조회 API","description":"사전결제를 상세 조회하는 API 입니다.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"사전결제 상세 조회 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryPrepareRetrieveResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[4] 사전결제 (DeliveryPrepare): 완료"],"security":[{"bearer":[]}]}},"/api/deliverys/cancels":{"post":{"operationId":"DeliveryCancelController_createPrepare","summary":"결제취소 생성 API","description":"결제취소를 생성하는 API 입니다.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDeliveryCancelDto"}}}},"responses":{"200":{"description":"결제취소 생성 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryCancelListResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[6] 결제취소 (DeliveryCancel): 진행중"],"security":[{"bearer":[]}]}},"/api/deliverys/cancels/{id}":{"get":{"operationId":"DeliveryCancelController_getDeliveryCancel","summary":"결제취소 상세 조회 API","description":"결제취소를 상세 조회하는 API 입니다.","parameters":[],"responses":{"200":{"description":"결제취소 상세 조회 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryCancelRetrieveResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[6] 결제취소 (DeliveryCancel): 진행중"],"security":[{"bearer":[]}]}},"/api/cards":{"post":{"operationId":"CardController_createCard","summary":"카드 생성 API","description":"카드 생성하는 API 입니다.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCardDto"}}}},"responses":{"200":{"description":"카드 생성 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardCreateResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[3] 카드 등록 (Card): 완료"],"security":[{"bearer":[]}]},"get":{"operationId":"CardController_getCards","summary":"카드 리스트 조회 API","description":"카드 리스트를 조회하는 API 입니다.","parameters":[],"responses":{"200":{"description":"카드 리스트 조회 성공시 결과 예시","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/Paginated"},{"properties":{"statusCode":{"type":"number","default":200},"message":{"type":"string","default":"리스트 조회를 완료하였습니다"},"data":{"type":"array","items":{"$ref":"#/components/schemas/CardEntity"}}}}]}}}},"401":{"description":"Unauthorized"}},"tags":["[3] 카드 등록 (Card): 완료"],"security":[{"bearer":[]}]}},"/api/cards/{id}":{"get":{"operationId":"CardController_getCard","summary":"카드 상세 조회 API","description":"카드 상세 조회하는 API 입니다.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"카드 상세 조회 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardRetrieveResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[3] 카드 등록 (Card): 완료"],"security":[{"bearer":[]}]},"patch":{"operationId":"CardController_partialUpdateCard","summary":"카드 부분 수정 API","description":"카드 부분 수정하는 API 입니다.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"카드 부분 수정 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardUpdateResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[3] 카드 등록 (Card): 완료"],"security":[{"bearer":[]}]},"put":{"operationId":"CardController_updateCard","summary":"카드 전체 수정 API","description":"카드 전체 수정하는 API 입니다.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"카드 전체 수정 성공시 결과 예시","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardUpdateResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["[3] 카드 등록 (Card): 완료"],"security":[{"bearer":[]}]}},"/health":{"get":{"operationId":"HealthCheckerController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}},"tags":["[0] 기본 (Default)"]}}},"info":{"title":"Delivery Admin API","description":"### REST\n\nRoutes is following REST standard (Richardson level 3)\n\n<details><summary>Detailed specification</summary>\n<p>\n\n**List:**\n  - `GET /<resources>/`\n    - Get the list of **<resources>** as admin\n  - `GET /user/<user_id>/<resources>/`\n    - Get the list of **<resources>** for a given **<user_id>**\n    - Output a **403** if logged user is not **<user_id>**\n\n**Detail:**\n  - `GET /<resources>/<resource_id>`\n    - Get the detail for **<resources>** of id **<resource_id>**\n    - Output a **404** if not found\n  - `GET /user/<user_id>/<resources>/<resource_id>`\n    - Get the list of **<resources>** for a given **user_id**\n    - Output a **404** if not found\n    - Output a **403** if:\n      - Logged user is not **<user_id>**\n      - The **<user_id>** have no access to **<resource_id>**\n\n**Creation / Edition / Replacement / Suppression:**\n  - `<METHOD>` is:\n    - **POST** for creation\n    - **PATCH** for update (one or more fields)\n    - **PUT** for replacement (all fields, not used)\n    - **DELETE** for suppression (all fields, not used)\n  - `<METHOD> /<resources>/<resource_id>`\n    - Create **<resources>** with id **<resource_id>** as admin\n    - Output a **400** if **<resource_id>** conflicts with existing **<resources>**\n  - `<METHOD> /user/<user_id>/<resources>/<resource_id>`\n    - Create **<resources>** with id **<resource_id>** as a given **user_id**\n    - Output a **409** if **<resource_id>** conflicts with existing **<resources>**\n    - Output a **403** if:\n      - Logged user is not **<user_id>**\n      - The **<user_id>** have no access to **<resource_id>**\n</p>\n</details>","version":"v1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"key":{"type":"string","description":"고유 키","default":"key-here"},"name":{"type":"string","description":"이름","default":"홍길동"},"phone":{"type":"string","description":"전화번호","default":"01040840660"},"email":{"type":"string","description":"이메일","default":"sample@test.com"},"address":{"type":"string","description":"주소","default":"서울 강남구"},"postcode":{"type":"string","description":"우편번호","default":"06325"}}},"UserTokenDto":{"type":"object","properties":{"id":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"key":{"type":"string","description":"고유 키","default":"key-here"},"name":{"type":"string","description":"이름","default":"홍길동"},"phone":{"type":"string","description":"전화번호","default":"01040840660"},"email":{"type":"string","description":"이메일","default":"sample@test.com"},"address":{"type":"string","description":"주소","default":"서울 강남구"},"postcode":{"type":"string","description":"우편번호","default":"06325"},"expiresIn":{"type":"number","default":3600},"accessToken":{"type":"string","default":"eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIxZTA2ZDQ1Zi03YzUzLTQzNTYtODlhOC02ZWQ2OTBiNzM1YjciLCJ0eXBlIjoiQUNDRVNTX1RPS0VOIiwicm9sZSI6IlVTRVIiLCJpYXQiOjE2NTM2NTA1MzJ9.Re1ElE6f5Hdw-YY5Gfto2Tj1jOGB5QkyitWrOyh6HH42X2htB8M8iyumpbYvnAbPHk1A0YwRZjrtMLaSvpZlkg"}},"required":["expiresIn","accessToken"]},"UserRegisteredResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":201},"message":{"type":"string","default":"유저 등록이 완료하였습니다."},"data":{"$ref":"#/components/schemas/UserTokenDto"}},"required":["statusCode","message","data"]},"UserLoginDto":{"type":"object","properties":{"id":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"}}},"UserLoginedResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":200},"message":{"type":"string","default":"유저 로그인을 완료하였습니다."},"data":{"$ref":"#/components/schemas/UserTokenDto"}},"required":["statusCode","message","data"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string","description":"이름","default":"홍길동"},"phone":{"type":"string","description":"전화번호","default":"01040840660"},"email":{"type":"string","description":"이메일","default":"sample@test.com"},"address":{"type":"string","description":"주소","default":"서울 강남구"},"postcode":{"type":"string","description":"우편번호","default":"06325"}}},"UserEntity":{"type":"object","properties":{"id":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"key":{"type":"string","description":"고유 키","default":"key-here"},"name":{"type":"string","description":"이름","default":"홍길동"},"phone":{"type":"string","description":"전화번호","default":"01040840660"},"email":{"type":"string","description":"이메일","default":"sample@test.com"},"address":{"type":"string","description":"주소","default":"서울 강남구"},"postcode":{"type":"string","description":"우편번호","default":"06325"}},"required":["id","key","name","phone","email","address","postcode"]},"UserUpdateResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":200},"message":{"type":"string","default":"수정을 완료하였습니다."},"data":{"$ref":"#/components/schemas/UserEntity"},"errors":{"type":"object"}},"required":["statusCode"]},"CreateDeliveryDto":{"type":"object","properties":{"deliveryPrepareId":{"type":"string","description":"유저 id","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"impUid":{"type":"string"}}},"DeliveryEntity":{"type":"object","properties":{"id":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"sellerId":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"buyerId":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"cardId":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"deliveryPrepareId":{"type":"string","description":"유저 id","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"impUid":{"type":"string"},"merchantUid":{"type":"string","description":"가맹점 거래 고유번호"},"payMethod":{"type":"string"},"channel":{"type":"string"},"pgProvider":{"type":"string"},"embPgProvider":{"type":"string"},"pgTid":{"type":"string"},"pgId":{"type":"string"},"escrow":{"type":"boolean"},"applyNum":{"type":"string"},"bankCode":{"type":"string"},"bankName":{"type":"string"},"cardCode":{"type":"string"},"cardName":{"type":"string"},"cardQuota":{"type":"number"},"cardNumber":{"type":"string"},"cardType":{"type":"string"},"vbankCode":{"type":"string"},"vbankName":{"type":"string"},"vbankNum":{"type":"string"},"vbankHolder":{"type":"string"},"vbankDate":{"type":"number"},"vbankIssuedAt":{"type":"number"},"name":{"type":"string"},"amount":{"type":"number"},"cancelAmount":{"type":"number"},"currency":{"type":"string"},"buyerName":{"type":"string"},"buyerEmail":{"type":"string"},"buyerTel":{"type":"string"},"buyerAddr":{"type":"string"},"buyerPostcode":{"type":"string"},"customData":{"type":"string"},"userAgent":{"type":"string"},"status":{"type":"string"},"startedAt":{"type":"number"},"paidAt":{"type":"number"},"failedAt":{"type":"number"},"cancelledAt":{"type":"number"},"failReason":{"type":"string"},"cancelReason":{"type":"string"},"receiptUrl":{"type":"string"},"cashReceiptIssued":{"type":"boolean"},"customerUid":{"type":"string"},"customerUidUsage":{"type":"string"}},"required":["id","sellerId","buyerId","cardId","deliveryPrepareId"]},"DeliveryCreateResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":201},"message":{"type":"string","default":"생성을 완료하였습니다."},"data":{"$ref":"#/components/schemas/DeliveryEntity"},"errors":{"type":"object"}},"required":["statusCode"]},"CreateCardDeliveryDto":{"type":"object","properties":{"sellerId":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"cardId":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"deliveryPrepareId":{"type":"string","description":"유저 id","default":"7e5f0280-db53-11ec-9d64-0242ac120002"}}},"DeliveryListResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":200},"message":{"type":"string","default":"리스트 조회를 완료하였습니다."},"data":{"items":{"type":"array"},"type":"array"},"errors":{"type":"object"}},"required":["statusCode"]},"DeliveryRetrieveResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":200},"message":{"type":"string","default":"상세 조회를 완료하였습니다."},"data":{"$ref":"#/components/schemas/DeliveryEntity"},"errors":{"type":"object"}},"required":["statusCode"]},"CreateDeliveryPrepareDto":{"type":"object","properties":{"sellerId":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"productId":{"type":"number","default":1},"amount":{"type":"number","default":1000}}},"DeliveryPrepareEntity":{"type":"object","properties":{"id":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"sellerId":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"buyerId":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"merchantUid":{"type":"string"},"productId":{"type":"number","default":1},"amount":{"type":"number","default":1000},"deliveryId":{"type":"string","description":"결제 id"}},"required":["id","sellerId","buyerId","merchantUid","productId","amount","deliveryId"]},"DeliveryPrepareCreateResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":201},"message":{"type":"string","default":"생성을 완료하였습니다."},"data":{"$ref":"#/components/schemas/DeliveryPrepareEntity"},"errors":{"type":"object"}},"required":["statusCode"]},"Paginated":{"type":"object","properties":{}},"DeliveryPrepareRetrieveResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":200},"message":{"type":"string","default":"상세 조회를 완료하였습니다."},"data":{"$ref":"#/components/schemas/DeliveryPrepareEntity"},"errors":{"type":"object"}},"required":["statusCode"]},"CreateDeliveryCancelDto":{"type":"object","properties":{"deliveryId":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"amount":{"type":"number","default":1000}}},"DeliveryCancelListResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":200},"message":{"type":"string","default":"리스트 조회를 완료하였습니다."},"data":{"items":{"type":"array"},"type":"array"},"errors":{"type":"object"}},"required":["statusCode"]},"DeliveryCancelEntity":{"type":"object","properties":{"id":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"deliveryId":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"amount":{"type":"number","default":1000}},"required":["id","deliveryId","amount"]},"DeliveryCancelRetrieveResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":200},"message":{"type":"string","default":"상세 조회를 완료하였습니다."},"data":{"$ref":"#/components/schemas/DeliveryCancelEntity"},"errors":{"type":"object"}},"required":["statusCode"]},"CreateCardDto":{"type":"object","properties":{"title":{"type":"string","description":"카드명"},"cardNumber":{"type":"string","description":"카드번호(dddd-dddd-dddd-dddd)","default":"1234-5678-1234-5678"},"expiryYear":{"type":"number","description":"카드 유효기간(YYYY)","default":2022},"expiryMonth":{"type":"number","description":"카드 유효기간 월(MM)","default":6},"birth":{"type":"string","description":"생년월일6자리(법인카드의 경우 사업자등록번호10자리)","default":"920426"},"pwd2digit":{"type":"string","description":"카드비밀번호 앞 2자리","default":"12"},"cvc":{"type":"string","description":"카드 인증번호(CVC)","default":"123"}}},"CardEntity":{"type":"object","properties":{"id":{"type":"string","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"userId":{"type":"string","description":"유저 id","default":"7e5f0280-db53-11ec-9d64-0242ac120002"},"title":{"type":"string","description":"카드명"},"customerUid":{"type":"string","description":"string 타입의 고객 고유번호.\n(결제에 사용된 카드정보를 빌링키 형태로 저장해두고 재결제에 사용하시려면 customerUid를 지정해주세요. /subscribe/deliverys/again, /subscribe/deliverys/schedule로 재결제를 진행하실 수 있습니다.)"},"cardNumber":{"type":"string","description":"카드번호(dddd-dddd-dddd-dddd)","default":"1234-5678-1234-5678"},"expiryYear":{"type":"number","description":"카드 유효기간(YYYY)","default":2022},"expiryMonth":{"type":"number","description":"카드 유효기간 월(MM)","default":6},"expiry":{"type":"string","description":"카드 유효기간(YYYY-MM)"},"birth":{"type":"string","description":"생년월일6자리(법인카드의 경우 사업자등록번호10자리)","default":"920426"},"pwd2digit":{"type":"string","description":"카드비밀번호 앞 2자리","default":"12"},"cvc":{"type":"string","description":"카드 인증번호(CVC)","default":"123"},"lastCardNumber":{"type":"string","description":"카드번호 마지막 4자리","default":"1234"},"cardName":{"type":"string","description":"카드명","default":"Card Company"},"pgId":{"type":"string","description":"PG ID","default":"PG ID"},"pgProvider":{"type":"string","description":"PG","default":"PG"}},"required":["id","userId","title","customerUid","cardNumber","expiryYear","expiryMonth","expiry","birth","pwd2digit","cvc","lastCardNumber","cardName","pgId","pgProvider"]},"CardCreateResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":201},"message":{"type":"string","default":"생성을 완료하였습니다."},"data":{"$ref":"#/components/schemas/CardEntity"},"errors":{"type":"object"}},"required":["statusCode"]},"CardRetrieveResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":200},"message":{"type":"string","default":"상세 조회를 완료하였습니다."},"data":{"$ref":"#/components/schemas/CardEntity"},"errors":{"type":"object"}},"required":["statusCode"]},"CardUpdateResponse":{"type":"object","properties":{"statusCode":{"type":"object","default":200},"message":{"type":"string","default":"수정을 완료하였습니다."},"data":{"$ref":"#/components/schemas/CardEntity"},"errors":{"type":"object"}},"required":["statusCode"]}}}}